svn同步更新

+++ cpct-service/src/main/java/com/zjtelcom/cpct/service/impl/grouping/TarGrpServiceImpl.java
@@ -106,15 +106,21 @@ public class TarGrpServiceImpl extends BaseService implements TarGrpService {
     @Override
     public Map<String, Object> copyTarGrp(Long tarGrpId,boolean isCopy) {
         Map<String,Object> result = new HashMap<>();
-        TarGrp tarGrp = tarGrpMapper.selectByPrimaryKey(tarGrpId);
-        if (tarGrp==null){
+//        TarGrp tarGrp = tarGrpMapper.selectByPrimaryKey(tarGrpId);
+//        if (tarGrp==null){
+//            result.put("resultCode", CODE_FAIL);
+//            result.put("resultMsg", "请选择下拉框运算类型");
+//            return result;
+//        }
+//        List<TarGrpCondition> conditionList = tarGrpConditionMapper.listTarGrpCondition(tarGrpId);
+//        TarGrpDetail detail = BeanUtil.create(tarGrp,new TarGrpDetail());
+//        detail.setTarGrpConditions(conditionList);
+        TarGrpDetail detail = (TarGrpDetail)redisUtils.get("TAR_GRP_"+tarGrpId);
+        if (detail==null){
             result.put("resultCode", CODE_FAIL);
-            result.put("resultMsg", "请选择下拉框运算类型");
+            result.put("resultMsg", "客户分群不存在");
             return result;
         }
-        List<TarGrpCondition> conditionList = tarGrpConditionMapper.listTarGrpCondition(tarGrpId);
-        TarGrpDetail detail = BeanUtil.create(tarGrp,new TarGrpDetail());
-        detail.setTarGrpConditions(conditionList);
         result = createTarGrp(detail,isCopy);
         return result;
     }
@@ -123,22 +129,22 @@ public class TarGrpServiceImpl extends BaseService implements TarGrpService {
     /**
      * 模板创建客户分群
      * @param templateId
-     * @return
+     * @return‘
      */
     @Override
     public Map<String, Object> createTarGrpByTemplateId(Long templateId) {
         Map<String, Object> result = new HashMap<>();
-        TarGrpTemplateDO template = tarGrpTemplateMapper.selectByPrimaryKey(templateId);
+        TarGrp template = tarGrpMapper.selectByPrimaryKey(templateId);
         if (template==null){
             result.put("resultCode", CODE_FAIL);
             result.put("resultMsg", "模板不存在");
             return result;
         }
-        List<TarGrpTemplateConditionDO> conditionDOList = tarGrpTemplateConditionMapper.selectByTarGrpTemplateId(templateId);
+        List<TarGrpCondition> conditionDOList = tarGrpConditionMapper.listTarGrpCondition(templateId);
 
         TarGrpDetail addVO = BeanUtil.create(template,new TarGrpDetail());
         List<TarGrpCondition> conditionAdd = new ArrayList<>();
-        for (TarGrpTemplateConditionDO conditionDO : conditionDOList){
+        for (TarGrpCondition conditionDO : conditionDOList){
             TarGrpCondition con = BeanUtil.create(conditionDO,new TarGrpCondition());
             conditionAdd.add(con);
         }
@@ -169,26 +175,34 @@ public class TarGrpServiceImpl extends BaseService implements TarGrpService {
         }
         tarGrpMapper.createTarGrp(tarGrp);
         List<TarGrpCondition> tarGrpConditions = tarGrpDetail.getTarGrpConditions();
-        for (final TarGrpCondition tarGrpCondition : tarGrpConditions) {
-            if (tarGrpCondition.getOperType()==null || tarGrpCondition.getOperType().equals("")){
-                maps.put("resultCode", CODE_FAIL);
-                maps.put("resultMsg", "请选择下拉框运算类型");
-                return maps;
+        List<TarGrpCondition> conditionList = new ArrayList<>();
+        if(tarGrpConditions!=null && tarGrpConditions.size()>0){
+            for (TarGrpCondition tarGrpCondition : tarGrpConditions) {
+                if (tarGrpCondition.getOperType()==null || tarGrpCondition.getOperType().equals("")){
+                    maps.put("resultCode", CODE_FAIL);
+                    maps.put("resultMsg", "请选择下拉框运算类型");
+                    return maps;
+                }
+                if (tarGrpCondition.getAreaIdList()!=null){
+                    area2RedisThread(tarGrp, tarGrpCondition);
+                }
+                tarGrpCondition.setConditionId(null);
+                tarGrpCondition.setLeftParamType(LeftParamType.LABEL.getErrorCode());//左参为注智标签
+                tarGrpCondition.setRightParamType(RightParamType.FIX_VALUE.getErrorCode());//右参为固定值
+                tarGrpCondition.setTarGrpId(tarGrp.getTarGrpId());
+                tarGrpCondition.setCreateDate(DateUtil.getCurrentTime());
+                tarGrpCondition.setUpdateDate(DateUtil.getCurrentTime());
+                tarGrpCondition.setStatusDate(DateUtil.getCurrentTime());
+                tarGrpCondition.setUpdateStaff(UserUtil.loginId());
+                tarGrpCondition.setCreateStaff(UserUtil.loginId());
+                conditionList.add(tarGrpCondition);
             }
-            if (tarGrpCondition.getAreaIdList()!=null){
-                area2RedisThread(tarGrp, tarGrpCondition);
-            }
-            tarGrpCondition.setLeftParamType(LeftParamType.LABEL.getErrorCode());//左参为注智标签
-            tarGrpCondition.setRightParamType(RightParamType.FIX_VALUE.getErrorCode());//右参为固定值
-            tarGrpCondition.setTarGrpId(tarGrp.getTarGrpId());
-            tarGrpCondition.setCreateDate(DateUtil.getCurrentTime());
-            tarGrpCondition.setUpdateDate(DateUtil.getCurrentTime());
-            tarGrpCondition.setStatusDate(DateUtil.getCurrentTime());
-            tarGrpCondition.setUpdateStaff(UserUtil.loginId());
-            tarGrpCondition.setCreateStaff(UserUtil.loginId());
-            tarGrpCondition.setStatusCd(CommonConstant.STATUSCD_EFFECTIVE);
-            tarGrpConditionMapper.insert(tarGrpCondition);
+            tarGrpConditionMapper.insertByBatch(conditionList);
         }
+        //数据加入redis
+        TarGrpDetail detail = BeanUtil.create(tarGrp,new TarGrpDetail());
+        detail.setTarGrpConditions(conditionList);
+        redisUtils.set("TAR_GRP_"+tarGrp.getTarGrpId(),detail);
         //插入客户分群条件
         maps.put("resultCode", CommonConstant.CODE_SUCCESS);
         maps.put("resultMsg", StringUtils.EMPTY);
@@ -282,7 +296,18 @@ public class TarGrpServiceImpl extends BaseService implements TarGrpService {
     public Map<String, Object> delTarGrpCondition(Long conditionId) {
         Map<String, Object> mapsT = new HashMap<>();
         try {
+            TarGrpCondition condition = tarGrpConditionMapper.selectByPrimaryKey(conditionId);
+            if (condition==null){
+                mapsT.put("resultCode", CODE_FAIL);
+                mapsT.put("resultMsg", ErrorCode.DEL_TAR_GRP_CONDITION_FAILURE.getErrorMsg());
+                return mapsT;
+            }
+            Long tarGrpId = condition.getTarGrpId();
             tarGrpConditionMapper.deleteByPrimaryKey(conditionId);
+            List<TarGrpCondition> conditionList = tarGrpConditionMapper.listTarGrpCondition(tarGrpId);
+            if (conditionList.isEmpty()){
+                tarGrpMapper.deleteByPrimaryKey(tarGrpId);
+            }
         } catch (Exception e) {
             mapsT.put("resultCode", CODE_FAIL);
             mapsT.put("resultMsg", ErrorCode.DEL_TAR_GRP_CONDITION_FAILURE.getErrorMsg());
@@ -323,6 +348,8 @@ public class TarGrpServiceImpl extends BaseService implements TarGrpService {
         tarGrp.setUpdateStaff(UserUtil.loginId());
         tarGrpMapper.modTarGrp(tarGrp);
         List<TarGrpCondition> tarGrpConditions = tarGrpDetail.getTarGrpConditions();
+        List<TarGrpCondition> insertConditions = new ArrayList<>();
+        List<TarGrpCondition> allCondition = new ArrayList<>();
         for (TarGrpCondition tarGrpCondition : tarGrpConditions) {
             TarGrpCondition tarGrpCondition1 = tarGrpConditionMapper.selectByPrimaryKey(tarGrpCondition.getConditionId());
             if (tarGrpCondition1 == null) {
@@ -331,25 +358,35 @@ public class TarGrpServiceImpl extends BaseService implements TarGrpService {
                     maps.put("resultMsg", "请选择下拉框运算类型");
                     return maps;
                 }
+                TarGrpCondition condition = BeanUtil.create(tarGrpCondition,new TarGrpCondition());
                 if (tarGrpCondition.getAreaIdList()!=null){
                     area2RedisThread(tarGrp, tarGrpCondition);
                 }
-                tarGrpCondition.setLeftParamType(LeftParamType.LABEL.getErrorCode());//左参为注智标签
-                tarGrpCondition.setRightParamType(RightParamType.FIX_VALUE.getErrorCode());//右参为固定值
-                tarGrpCondition.setTarGrpId(tarGrp.getTarGrpId());
-                tarGrpCondition.setUpdateDate(DateUtil.getCurrentTime());
-                tarGrpCondition.setCreateDate(DateUtil.getCurrentTime());
-                tarGrpCondition.setStatusDate(DateUtil.getCurrentTime());
-                tarGrpCondition.setUpdateStaff(UserUtil.loginId());
-                tarGrpCondition.setCreateStaff(UserUtil.loginId());
-                tarGrpCondition.setStatusCd(CommonConstant.STATUSCD_EFFECTIVE);
-                tarGrpConditionMapper.insert(tarGrpCondition);
+                condition.setLeftParamType(LeftParamType.LABEL.getErrorCode());//左参为注智标签
+                condition.setRightParamType(RightParamType.FIX_VALUE.getErrorCode());//右参为固定值
+                condition.setTarGrpId(tarGrp.getTarGrpId());
+                condition.setUpdateDate(DateUtil.getCurrentTime());
+                condition.setCreateDate(DateUtil.getCurrentTime());
+                condition.setStatusDate(DateUtil.getCurrentTime());
+                condition.setUpdateStaff(UserUtil.loginId());
+                condition.setCreateStaff(UserUtil.loginId());
+                insertConditions.add(condition);
             } else {
                 tarGrpCondition.setUpdateDate(DateUtil.getCurrentTime());
                 tarGrpCondition.setUpdateStaff(UserUtil.loginId());
                 tarGrpConditionMapper.modTarGrpCondition(tarGrpCondition);
+                allCondition.add(tarGrpCondition);
             }
         }
+        if (!insertConditions.isEmpty()){
+            tarGrpConditionMapper.insertByBatch(insertConditions);
+        }
+        allCondition.addAll(insertConditions);
+        //更新redis分群数据
+        TarGrpDetail detail = BeanUtil.create(tarGrp,new TarGrpDetail());
+        detail.setTarGrpConditions(allCondition);
+        redisUtils.set("TAR_GRP_"+tarGrp.getTarGrpId(),detail);
+
         maps.put("resultCode", CommonConstant.CODE_SUCCESS);
         maps.put("resultMsg", StringUtils.EMPTY);
         return maps;
@@ -402,7 +439,7 @@ public class TarGrpServiceImpl extends BaseService implements TarGrpService {
      * 获取目标分群条件信息
      */
     @Override
-    public Map<String, Object> listTarGrpCondition(Long tarGrpId) throws Exception {
+    public Map<String, Object> listTarGrpCondition(Long tarGrpId){
         Map<String, Object> maps = new HashMap<>();
         if (tarGrpId==null){
             maps.put("resultCode", CODE_FAIL);
@@ -418,7 +455,8 @@ public class TarGrpServiceImpl extends BaseService implements TarGrpService {
         for (TarGrpCondition tarGrpCondition : listTarGrpCondition) {
             List<OperatorDetail> operatorList = new ArrayList<>();
             TarGrpConditionVO tarGrpConditionVO = new TarGrpConditionVO();
-            CopyPropertiesUtil.copyBean2Bean(tarGrpConditionVO, tarGrpCondition);
+            BeanUtil.copy(tarGrpCondition,tarGrpConditionVO);
+//            CopyPropertiesUtil.copyBean2Bean(tarGrpConditionVO, tarGrpCondition);
             //塞入左参中文名
             Label label = injectionLabelMapper.selectByPrimaryKey(Long.valueOf(tarGrpConditionVO.getLeftParam()));
             if (label==null){
@@ -479,6 +517,7 @@ public class TarGrpServiceImpl extends BaseService implements TarGrpService {
         maps.put("resultCode", CommonConstant.CODE_SUCCESS);
         maps.put("resultMsg", StringUtils.EMPTY);
         maps.put("listTarGrpCondition", grpConditionList);
+        maps.put("conditionList",listTarGrpCondition);
         return maps;
     }
